Refactor package structure and add Docker/README setup.
Reorganized the model response classes under a new versioned "rest.models.v2.responses" package to better align with API versioning. Added a Dockerfile for containerized test execution and a README file detailing project structure, build, and test instructions. Adjusted Maven compiler version and enhanced configuration with Javadoc comments.
