--- src/main/java/org/truenvy/dogs/model/response/breeds/Data.java
@@ -1,7 +1,0 @@
-package org.truenvy.dogs.model.response.breeds;
-
-public record Data(
-	String id,
-	String type
-) {
-}
