--- src/main/java/org/truenvy/dogs/model/response/breeds/Group.java
@@ -1,6 +1,0 @@
-package org.truenvy.dogs.model.response.breeds;
-
-public record Group(
-	Data data
-) {
-}
