--- src/main/java/org/truenvy/dogs/model/response/breeds/Life.java
@@ -1,7 +1,0 @@
-package org.truenvy.dogs.model.response.breeds;
-
-public record Life(
-	Integer min,
-	Integer max
-) {
-}
