--- src/main/java/org/truenvy/dogs/model/response/facts/Attributes.java
@@ -1,5 +1,0 @@
-package org.truenvy.dogs.model.response.facts;
-
-public record Attributes(
-	String body
-) {}
