+++ src/main/java/org/truenvy/dogs/rest/models/v2/responses/breeds/Data.java
@@ -1,0 +1,7 @@
+package org.truenvy.dogs.rest.models.v2.responses.breeds;
+
+public record Data(
+	String id,
+	String type
+) {
+}
