+++ src/main/java/org/truenvy/dogs/rest/models/v2/responses/breeds/FemaleWeight.java
@@ -1,4 +1,4 @@
-package org.truenvy.dogs.model.response.breeds;
+package org.truenvy.dogs.rest.models.v2.responses.breeds;
 
 public record FemaleWeight(
 	Integer min,
