+++ src/main/java/org/truenvy/dogs/rest/models/v2/responses/breeds/Group.java
@@ -1,0 +1,6 @@
+package org.truenvy.dogs.rest.models.v2.responses.breeds;
+
+public record Group(
+	Data data
+) {
+}
