+++ src/main/java/org/truenvy/dogs/rest/models/v2/responses/breeds/Life.java
@@ -1,0 +1,7 @@
+package org.truenvy.dogs.rest.models.v2.responses.breeds;
+
+public record Life(
+	Integer min,
+	Integer max
+) {
+}
