+++ src/main/java/org/truenvy/dogs/rest/models/v2/responses/breeds/Relationships.java
@@ -1,0 +1,6 @@
+package org.truenvy.dogs.rest.models.v2.responses.breeds;
+
+public record Relationships(
+	Group group
+) {
+}
