+++ src/main/java/org/truenvy/dogs/rest/models/v2/responses/facts/Attributes.java
@@ -1,0 +1,5 @@
+package org.truenvy.dogs.rest.models.v2.responses.facts;
+
+public record Attributes(
+	String body
+) {}
