+++ src/main/java/org/truenvy/dogs/rest/models/v2/responses/facts/FactsResponse.java
@@ -1,4 +1,4 @@
-package org.truenvy.dogs.model.response.facts;
+package org.truenvy.dogs.rest.models.v2.responses.facts;
 
 import java.util.List;
